Refactor chatbot wrapper to avoid duplication
