+++ src/components/ui/chatbot/constants.ts
@@ -1,0 +1,1 @@
+export const CHATBOT_BUTTON_DRAG_HANDLE_CLASS_NAME = 'chatbot-drag-handle';
